refactor(agents): move chat internals under server.agents
