+++ src/phoenix/server/agents/tools/__init__.py
@@ -1,0 +1,3 @@
+from phoenix.server.agents.tools.registry import resolve_contextual_tools
+
+__all__ = ["resolve_contextual_tools"]
