--- src/phoenix/server/api/routers/chat_tools/__init__.py
@@ -1,3 +1,0 @@
-from phoenix.server.api.routers.chat_tools.registry import resolve_contextual_tools
-
-__all__ = ["resolve_contextual_tools"]
